Move timer into separate package
Cleaning up the utilities
