+++ src/main/java/org/frontear/infinity/modules/impl/AutoClicker.java
@@ -3,8 +3,8 @@ package org.frontear.infinity.modules.impl;
 import net.minecraft.client.entity.EntityPlayerSP;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 import net.minecraftforge.fml.common.gameevent.TickEvent;
-import org.frontear.framework.utils.TimeUnit;
-import org.frontear.framework.utils.Timer;
+import org.frontear.framework.utils.time.TimeUnit;
+import org.frontear.framework.utils.time.Timer;
 import org.frontear.infinity.modules.Category;
 import org.frontear.infinity.modules.Module;
 import org.lwjgl.input.Keyboard;
