+++ src/main/java/org/frontear/infinity/modules/impl/Breadcrumbs.java
@@ -5,7 +5,7 @@ import net.minecraft.client.entity.EntityPlayerSP;
 import net.minecraft.util.Vec3;
 import net.minecraftforge.client.event.RenderWorldLastEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import org.frontear.framework.utils.Timer;
+import org.frontear.framework.utils.time.Timer;
 import org.frontear.infinity.events.entity.UpdateEvent;
 import org.frontear.infinity.modules.Category;
 import org.frontear.infinity.modules.Module;
